nekodigi/streamlit/streamlit-app/requirements.txt
streamlit>=1.28
openai>=1.0
Pillow>=10.0
